feat: create rpqmatrix submodule
